refactor: move model to separate folder
